feat: refactored user profile page
