+++ opal-ui/src/router/routes.ts
@@ -15,7 +15,7 @@ const routes: RouteRecordRaw[] = [
       { path: '/admin/settings', component: () => import('pages/AdminConfigurationPage.vue') },
       { path: '/admin/databases', component: () => import('pages/AdminDatabasesPage.vue') },
       { path: '/admin/profiles', component: () => import('pages/AdminProfilesPage.vue') },
-      { path: '/admin/profile/:principal/permissions', component: () => import('pages/AdminProfileAclsPage.vue') },
+      { path: '/admin/profile/:principal', component: () => import('pages/AdminProfilePage.vue') },
       { path: '/admin/idproviders', component: () => import('pages/AdminIdentityProvidersPage.vue') },
       { path: '/admin/plugins', component: () => import('pages/AdminPluginsPage.vue') },
       { path: '/admin/apps', component: () => import('pages/AdminAppsPage.vue') },
